+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 ---
 
-1. StaggeredGridLayoutManager 注意事项：
-[Android 关于RecyclerView瀑布流显示图片时Item切换、闪烁等问题优化](http://www.jianshu.com/p/81e088000ba6)
+1. [Android 关于RecyclerView瀑布流显示图片时Item切换、闪烁等问题优化](http://www.jianshu.com/p/81e088000ba6)
+
+2. [Android Material Design之CoordinatorLayout+AppBarLayout实现上滑隐藏ToolBar](http://blog.csdn.net/jdsjlzx/article/details/50421349)